🚚 Move generator form tamplate closer to its implementations
